Zip Validation
Provide two methods of CRC32 validation, ether reading the bytes
back out of the zip and recalcuating the CRC32 (bytes) or using the
CRC32 value from the Central Directory of the ZIP file.

The Central Directory valiation method must be used for ZIP files
that contain uncompressed files because the AsyncZip crate will
panic when decmpressing these files using the stream ZipReader.