feat/HSWE-3-zip-archive-validation (#18)
Added a crc32 validation type to the validation cli. The operator can choose between bytes and central-directory.
bytes will calculate the crc32 by reading the bytes of the entry and recalculating the digest. central-directory will
ready the crc32 value out of the central directory of the ZIP.

Added S3 AysncSeek and AsyncRead implementations for S3Object.

Changed the crc32 implementation from crc to crc32fast